fix: resolve all CI pipeline blocking issues and achieve zero-error build
This commit addresses all major compilation errors preventing the CI pipeline:

- Fixed CS0029 return type conversion errors in Apple Neural Engine async methods
- Fixed CS1503 platform interop argument conversion errors in Vulkan/ROCm platforms
- Fixed CS1061 missing method errors by correcting GetAsArray() to GetAsArray1D()
- Fixed CS8385 fixed statement syntax errors in Intel NPU by using proper unsafe pointers
- Fixed CS0315 type conversion errors in FFT accelerators with AsLinearView() calls
- Fixed CS0246 missing type errors in Intel AMX by adding proper using directives
- Fixed void* to IntPtr conversions in ROCmNative matrix operations
- Fixed null\! parameter issues in Vulkan API calls
- Fixed conditional compilation syntax in IntelAMXDevice for cross-platform builds

All compilation errors have been resolved achieving a zero-error build state.

+++ Src/ILGPU/Runtime/ROCm/ROCmMemoryBuffer.cs
@@ -116,7 +116,7 @@ namespace ILGPU.Runtime.ROCm
             else
             {
                 // Synchronous fallback
-                SetMemoryToValue(targetView.Index, value, targetView.Length);
+                SetMemoryToValue((nint)targetView.Index, value, (nint)targetView.Length);
             }
         }
 
@@ -148,7 +148,7 @@ namespace ILGPU.Runtime.ROCm
                     try
                     {
                         var result = ROCmNative.Memcpy(
-                            targetPtr, sourcePtr, (nuint)targetView.Length, 
+                            targetPtr, sourcePtr, (ulong)targetView.Length, 
                             HipMemcpyKind.HostToDevice);
                         ROCmException.ThrowIfFailed(result);
                     }
@@ -204,7 +204,7 @@ namespace ILGPU.Runtime.ROCm
                     try
                     {
                         var result = ROCmNative.Memcpy(
-                            targetPtr, sourcePtr, (nuint)sourceView.Length,
+                            targetPtr, sourcePtr, (ulong)sourceView.Length,
                             HipMemcpyKind.DeviceToHost);
                         ROCmException.ThrowIfFailed(result);
                     }
